Nunchuk Pitch + Roll Angle Updates (#14)
* Changed output to float

No reason to do floating point calculations and return an int

* Get roll full range

(1023 max - 1 min) / 2 = 511

* Get pitch full range

(1023 max - 1 min) / 2 = 511

* Inverted pitch angle

+++ src/Nunchuck.h
@@ -23,8 +23,8 @@ public:
     int getJoyX();
     int getJoyY();
 
-    int getRollAngle();
-    int getPitchAngle();
+    float getRollAngle();
+    float getPitchAngle();
     int getAccelX();
     int getAccelY();
     int getAccelZ();
